be/requirements.txt
attrs==22.1.0
boto3==1.28.61
botocore==1.31.61
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
Flask==2.2.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
PyJWT==2.8.0
pyparsing==3.0.9
pyrsistent==0.19.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==1.4.42
stripe==6.7.0
swagger-ui-bundle==0.0.9
typing_extensions==4.8.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.17.0
